requirements.txt
youtube_search==2.1.0
spotipy==2.19.0
youtube_dl==2021.6.6
requests==2.22.0
python-dotenv==0.19.0
